Update plugin developer guide
This information has been written based on what I've learned while trying to develop an Elm Package plugin.

I've tried to document it as accurately as I can based on my reading of the source code and my reading of existing plugins, as well as what I've observed in practice. 

This obviously isn't an exhaustive guide or tutorial, but hopefully it's helpful to other people starting out writing plugins, and can be the base of more complete docs later.