docs: improve the docs navigation
Links to the previous and next articles were added in all articles to improve the linear reading. Currently, the reader must back to the index to select the next article. After this change, the reader will jump to the next article by clicking in a link at the end of the page.