feat!: support multiple git tags and publish major and minor version tag
This adds support for multiple git tags to be specified as input. The user
can now use three new variables: `major`, `minor`, and `patch`.

The default tags aren't only `v${version}` anymore, but `v${major}.${minor}`
and `v${major}`.

This new pattern helps to spread the good practice of creating a git tag
with the major version number so the users can quickly identify what's the
revision of the latest major (or minor) release.

BREAKING CHANGE: the CLI parameter `--tag-format` was renamed to
`--tag-formats` and the `gitTag` variable in the `release`s objects from the
JavaScript API was renamed to `gitTags`, with the `Array` type rather than
`String`. For those relying on the `gitTag` variable, you should now use
`gitTags[0]`.

Fixes #1515