fix(version): use configured prerelease id
When `branch.prerelease` is set to a String, it's expected that this
will be used as the prerelease for the version. However, if a previous
version using this branch name is found in the list of tags, the logic
here will always choose to use that branch name as the prerelease ID and
ignore your configuration. This makes it effectively impossible to
change a prerelease ID to something different than the branch name if
you make a mistake (which is very easy to do since this is mostly
undocumented behavior).

Fixes #2177