feat: make printing Markdown to the Terminal optional
This commit adds a new configuration option called 'marked'.
When set to true, the output of Semantic Release would print Markdown to the
terminal, whenever supported, using the 'marked-terminal' library as usual.

https://www.npmjs.com/package/marked-terminal

When set to false, the output will be printed in plain text.
This feature is particularly useful when the output or the command will be
consumed by another Markdown renderer, like GitHub.

This addresses the following issue.
https://github.com/semantic-release/semantic-release/issues/2191