fix: remove marked deprecation warning
This commit disables to flags that are deprecated in the 'marked-terminal' as
pointed in the following issue.

https://github.com/markedjs/marked/issues/2793

```
marked(): mangle parameter is enabled by default, but is deprecated since
version 5.0.0, and will be removed in the future. To clear this warning, install
https://www.npmjs.com/package/marked-mangle, or disable by setting
`{mangle: false}`.

marked(): headerIds and headerPrefix parameters enabled by default, but are
deprecated since version 5.0.0, and will be removed in the future. To clear this
warning, install  https://www.npmjs.com/package/marked-gfm-heading-id, or
disable by setting `{headerIds: false}`.
```