Allow running semantic-release with Bun
I've tested this and it works fine, and semantic-release can run with no issues in a Bun-only environment (no Node).

However, Bun reports its Node version as 18.15. So, when you try to run semantic-release without this change, you get this error:

```
 ~/Code/ bunx --bun semantic-release --version
[semantic-release]: node version ^18.17 || >=20.6.1 is required. Found v18.15.0.

See https://github.com/semantic-release/semantic-release/blob/master/docs/support/node-version.md for more details and solutions.
```

Given that semantic-release works fine with Bun, we should not exit if we are running in the Bun runtime.