perf: read `package.json` directly
We can switch to `empathic` which is much faster at traversal.

On top of that, we don't actually rely on any of the normalisation going
on under the hood in the current dependency. So we can just `readFile`
the `package.json` directly.

We also don't need a plain object check, since we know it came from
`JSON.parse` (can't be anything else).